@@ -26,34 +26,39 @@
   "main": "es5/react-pixi-commonjs.js",
   "dependencies": {
     "brfs": "^1.4.1",
-    "fbjs": "^0.2.1"
+    "fbjs": "^0.7.0",
+    "object-assign": "^4.0.1"
   },
   "peerDependencies": {
     "pixi.js": "~3.0.7",
-    "react": "^0.14.2",
-    "react-dom": "^0.14.2"
+    "react": "^15.0.0",
+    "react-dom": "^15.0.0"
   },
   "devDependencies": {
-    "babel-core": "^5.8.23",
-    "babel-eslint": "^4.1.1",
-    "babel-loader": "^5.3.2",
-    "babel-runtime": "^5.8.20",
-    "eslint": "^1.4.0",
+    "babel-core": "^6.0.0",
+    "babel-eslint": "^6.0.0",
+    "babel-loader": "^6.0.0",
+    "babel-plugin-transform-runtime": "^6.6.0",
+    "babel-preset-es2015": "^6.0.0",
+    "babel-preset-react": "^6.0.0",
+    "babel-preset-stage-2": "^6.0.0",
+    "babel-runtime": "^6.6.1",
+    "eslint": "^2.4.0",
     "expose-loader": "^0.7.0",
     "json-loader": "^0.5.3",
     "karma": "~0.13.0",
     "karma-firefox-launcher": "~0.1.3",
     "karma-jasmine": "~0.3.2",
-    "lodash": "^3.3.1",
+    "lodash": "^4.6.1",
     "node-static": "~0.7.3",
     "pixi.js": "~3.0.7",
-    "react": "^0.14.2",
-    "react-dom": "^0.14.2",
-    "resemblejs": "~1.2.0",
-    "rimraf": "~2.4.0",
+    "react": "^15.0.0",
+    "react-dom": "^15.0.0",
+    "resemblejs": "~2.2.0",
+    "rimraf": "~2.5.0",
     "transform-loader": "^0.2.3",
     "webpack": "^1.12.1",
-    "webpack-dev-server": "^1.10.1"
+    "webpack-dev-server": "^1.14.1"
   },
   "scripts": {
     "build": "webpack",
@@ -82,7 +87,8 @@
     "rules": {
       "strict": [
         0
-      ]
+      ],
+      "eqeqeq": 2
     }
   },
   "engines": {
