@@ -24,10 +24,9 @@ module.exports = {
 	loader: 'babel',
 	include: path.join(__dirname, 'src'),
 	query: {
-	  // When generating a standalone library, this makes sure to
-	  // use babel-runtime to wrap our code and
-	  // avoid global polyfills.
-	  optional: ['runtime']
+          cacheDirectory: true,
+          presets: ['es2015', 'stage-2'],
+          plugins: ['transform-runtime']
 	}
       },
       {
